streamlit-app/requirements.txt
streamlit>=1.28.0
pyyaml>=6.0
pandas>=2.0.0
plotly>=5.15.0
